# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Trend Terminal - WSB Meme Stock Investigator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PostSentiment = typeof PostSentiment[keyof typeof PostSentiment];


export const PostSentiment = {
  Positive: 'Positive',
  Negative: 'Negative',
  Neutral: 'Neutral',
} as const;

export interface Post {
  post_id: string;
  title: string;
  /** @nullable */
  flair?: string | null;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  num_comments?: number | null;
  /** @nullable */
  date_utc?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  ticker?: string | null;
  sentiment: PostSentiment;
  /** @nullable */
  classification_method?: string | null;
}

export interface PostList {
  posts: Post[];
  total: number;
  offset: number;
  limit: number;
}

export interface SentimentCount {
  count: number;
  pct: number;
}

export interface SentimentSummary {
  Positive: SentimentCount;
  Negative: SentimentCount;
  Neutral: SentimentCount;
  total: number;
}

export interface TickerSummary {
  ticker: string;
  mention_count: number;
  positive_pct: number;
  negative_pct: number;
  neutral_pct: number;
  dominant_sentiment: string;
}

export interface SentimentDataPoint {
  date: string;
  Positive: number;
  Negative: number;
  Neutral: number;
}

export interface TickerPrice {
  ticker: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  current_price?: number | null;
  /** @nullable */
  price_change_pct?: number | null;
  /** @nullable */
  market_cap?: number | null;
  /** @nullable */
  week_52_high?: number | null;
  /** @nullable */
  week_52_low?: number | null;
  /** @nullable */
  sector?: string | null;
}

export interface StockInfo {
  ticker: string;
  name: string;
  sector: string;
  industry: string;
  current_price: number;
  previous_close: number;
  price_change: number;
  price_change_pct: number;
  /** @nullable */
  market_cap?: number | null;
  /** @nullable */
  week_52_high?: number | null;
  /** @nullable */
  week_52_low?: number | null;
  description: string;
  /** @nullable */
  error?: string | null;
}

export interface PricePoint {
  date: string;
  /** @nullable */
  open?: number | null;
  /** @nullable */
  high?: number | null;
  /** @nullable */
  low?: number | null;
  /** @nullable */
  close?: number | null;
  /** @nullable */
  volume?: number | null;
}

export interface ClassifyInput {
  text: string;
  /** @nullable */
  flair?: string | null;
}

export type ClassifyResultSentiment = typeof ClassifyResultSentiment[keyof typeof ClassifyResultSentiment];


export const ClassifyResultSentiment = {
  Positive: 'Positive',
  Negative: 'Negative',
  Neutral: 'Neutral',
} as const;

export interface ClassifyResult {
  sentiment: ClassifyResultSentiment;
  method: string;
  /** @nullable */
  ticker?: string | null;
  confidence: string;
}

export type ListPostsParams = {
sentiment?: ListPostsSentiment;
limit?: number;
offset?: number;
search?: string;
};

export type ListPostsSentiment = typeof ListPostsSentiment[keyof typeof ListPostsSentiment];


export const ListPostsSentiment = {
  Positive: 'Positive',
  Negative: 'Negative',
  Neutral: 'Neutral',
} as const;

export type GetTopTickersParams = {
limit?: number;
};

export type GetRecentActivityParams = {
limit?: number;
};

export type FetchStockHistoryParams = {
period?: FetchStockHistoryPeriod;
};

export type FetchStockHistoryPeriod = typeof FetchStockHistoryPeriod[keyof typeof FetchStockHistoryPeriod];


export const FetchStockHistoryPeriod = {
  '1wk': '1wk',
  '1mo': '1mo',
  '3mo': '3mo',
  '6mo': '6mo',
  '1y': '1y',
} as const;