# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: The Trend Terminal - WSB Meme Stock Investigator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stocks/{ticker}:
    get:
      operationId: getStockInfo
      tags: [stocks]
      summary: Get live stock info
      description: Returns live stock price and metadata for a ticker
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockInfo"
  /stocks/{ticker}/history:
    get:
      operationId: fetchStockHistory
      tags: [stocks]
      summary: Get price history
      description: Returns OHLCV price history for a ticker
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          schema:
            type: string
            default: 1mo
            enum: [1wk, 1mo, 3mo, 6mo, 1y]
      responses:
        "200":
          description: Price history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PricePoint"
...
components:
  schemas:
...
    StockInfo:
      type: object
      properties:
        ticker:
          type: string
        name:
          type: string
        sector:
          type: string
        industry:
          type: string
        current_price:
          type: number
        previous_close:
          type: number
        price_change:
          type: number
        price_change_pct:
          type: number
        market_cap:
          type: ["number", "null"]
        week_52_high:
          type: ["number", "null"]
        week_52_low:
          type: ["number", "null"]
        description:
          type: string
        error:
          type: ["string", "null"]
      required:
        - ticker
        - name
        - sector
        - industry
        - current_price
        - previous_close
        - price_change
        - price_change_pct
        - description
    PricePoint:
      type: object
      properties:
        date:
          type: string
        open:
          type: ["number", "null"]
        high:
          type: ["number", "null"]
        low:
          type: ["number", "null"]
        close:
          type: ["number", "null"]
        volume:
          type: ["number", "null"]
      required:
        - date
...