# lib/api-zod/src/generated/types/pricePoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Trend Terminal - WSB Meme Stock Investigator API
 * OpenAPI spec version: 0.1.0
 */

export interface PricePoint {
  date: string;
  /** @nullable */
  open?: number | null;
  /** @nullable */
  high?: number | null;
  /** @nullable */
  low?: number | null;
  /** @nullable */
  close?: number | null;
  /** @nullable */
  volume?: number | null;
}