# lib/api-zod/src/generated/types/stockInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Trend Terminal - WSB Meme Stock Investigator API
 * OpenAPI spec version: 0.1.0
 */

export interface StockInfo {
  ticker: string;
  name: string;
  sector: string;
  industry: string;
  current_price: number;
  previous_close: number;
  price_change: number;
  price_change_pct: number;
  /** @nullable */
  market_cap?: number | null;
  /** @nullable */
  week_52_high?: number | null;
  /** @nullable */
  week_52_low?: number | null;
  description: string;
  /** @nullable */
  error?: string | null;
}